rag_challenge/requirements.txt
pymupdf>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
requests>=2.32.0
tqdm>=4.66.0
